Move Recommendation buttons to their own files
